Updating the script based on feedback
I got two main points of feedback:

First was to give functions more specific names, even if they are 4-5 words long.

Second was an article about best practices in Python that I read and tried to change the whole structure of the script to reflect these practices.

As a side-note I found it hard to get people to give specific feedback on the code itself. A lot of it was just "looks good" and I really don't think my code is flawless so from now I will make comments with specific questions on parts I want feedback on. Of course I am still open for feedback on any other part!

Hopefully this will make it easier for you guys to give the feedback of which I am i desperate need.