finished slotmachine
It's super unfair (so it probably work like it should)

The base of this makes it so that if you bet on 1 line, you only bet on the first one, that could be a thing to change in the future

Probably change location of "press enter to play" and make it "press enter to spin again, type q to quit"

I have aleady made one change to the script which is that instead of copying all the code from each input function with slight variations I made one that checks for valid values and prints a string based on what I send though the function. it seems much less repetitive.